GHFunctionProject/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions==1.24.0
requests==2.32.4
parse==1.20.1f
packaging==25.0
pip-audit==2.10.0